src/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
